+++ src/davinci_crawling/example/bovespa/crawling_parts/crawl_listed_companies.py
@@ -1,27 +1,23 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2019 BuildGroup Data Services Inc.
 
+import itertools
 import logging
 import re
-import itertools
 import traceback
-
-from dateutil.parser import parse as date_parse
-
-from bs4 import BeautifulSoup
-
 from multiprocessing.pool import ThreadPool as Pool
 
-from selenium.webdriver.support.wait import WebDriverWait
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
-
+from bs4 import BeautifulSoup
+from dateutil.parser import parse as date_parse
 from davinci_crawling.example.bovespa import BOVESPA_CRAWLER
 from davinci_crawling.example.bovespa.models import \
     BovespaCompany, SITUATION_CANCELLED, SITUATION_GRANTED
-from davinci_crawling.crawling_throttle import Throttle
+from davinci_crawling.throttle.throttle import Throttle
 from davinci_crawling.utils import setup_cassandra_object_mapper, \
     CrawlersRegistry
+from selenium.webdriver.common.by import By
+from selenium.webdriver.support import expected_conditions as EC
+from selenium.webdriver.support.wait import WebDriverWait
 
 ALPHABET_LIST = list(map(chr, range(65, 91)))
 NUMBERS_LIST = list(range(0, 10))
@@ -39,7 +35,7 @@ _logger = logging.getLogger(
     format(BOVESPA_CRAWLER))
 
 
-@Throttle(minutes=1, rate=50, max_tokens=50)
+@Throttle(crawler_name=BOVESPA_CRAWLER, minutes=1, rate=50, max_tokens=50)
 def update_listed_companies(letter, options):
     """
     :param letter:
@@ -157,6 +153,10 @@ def crawl_listed_companies(options, workers_num=10):
         func_params = []
         companies_initials = options.get("crawling_initials",
                                          COMPANIES_LISTING_SEARCHER_LETTERS)
+
+        if not companies_initials:
+            companies_initials = COMPANIES_LISTING_SEARCHER_LETTERS
+
         for letter in companies_initials:
             func_params.append([letter, options])